[SASS-5228] Rename branch. Rename IncomeFromPensionsRedirects
